Refactor 'startEditors' method
