+++ browser/src/Services/Language/index.ts
@@ -2,7 +2,6 @@ export * from "./addInsertModeLanguageFunctionality"
 export * from "./CodeAction"
 export * from "./Definition"
 export * from "./DefinitionRequestor"
-export * from "./Diagnostics"
 export * from "./Edits"
 export * from "./FindAllReferences"
 export * from "./Formatting"
